Feature/pan channel

+++ src/csurf_faderport_8/csurf_fp_8_menu_manager.cpp
@@ -145,7 +145,7 @@ public:
     void HandleSelectClick(int index) override
     {
         (void)index;
-        int max_items = level == 0 ? menu_items.size() : static_cast<int>(menu_options.at(option[0]).size());
+        int max_items = level == 0 ? (int)menu_items.size() : static_cast<int>(menu_options.at(option[0]).size());
         if (level > 0 && option[1] == max_items)
         {
             level = 0;
@@ -183,7 +183,7 @@ public:
     {
         (void)index;
         (void)value;
-        int max_items = level == 0 ? menu_items.size() : static_cast<int>(menu_options.at(option[0]).size());
+        int max_items = level == 0 ? (int)menu_items.size() : static_cast<int>(menu_options.at(option[0]).size());
         option[level] = minmax(0, option[level] + 1, max_items - (level == 0 ? 1 : 0));
     }
 
@@ -192,7 +192,7 @@ public:
     {
         (void)index;
         (void)value;
-        int max_items = level == 0 ? menu_items.size() : static_cast<int>(menu_options.at(option[0]).size());
+        int max_items = level == 0 ? (int)menu_items.size() : static_cast<int>(menu_options.at(option[0]).size());
         option[level] = minmax(0, option[level] - 1, max_items - (level == 0 ? 1 : 0));
     }
